bench: share csharp benchmark packaging
